refactor: Service 계층 분리 #8
